Create hive_query_csv_export.sh
a general bash shell template for running a series or batch of Hive queries and storing the results in .csv files for analysis. this helpful for someone who might want to run a series of queries overnight and have results to analyze during the next business day.

future work: figure out how to schedule the running of this script with `cron` and also enable email notification of a job's completion.